niffler/requirements.txt
djangorestframework==3.9.2
django-rest-swagger==2.2.0
django-filter==2.1.0
Django==2.1.7
Pygments==2.3.1
Markdown==3.1
coreapi==2.3.3
psycopg2-binary==2.7.7
dj-database-url==0.5.0
gunicorn==19.9.0
whitenoise==4.1.2
PyYAML==5.1
django-cors-headers==3.0.2
django-partial-date==1.2.2
Pillow==4.3.0
pycryptodomex==3.7.2 
python-alipay-sdk==1.10.1
python-dateutil==2.6.0
